+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## dev
 
+* hal: added tests for `/api/hal/users` and `/api/hal/users/:id`
 * hal: added owner of notes
 * hal: added hal users
 * hal: extra folder for hal api modules